Textured Globe
The globe is now textured. Need to move implementations out of header
files so it doesn't take so long to compile small changes. Things are
currently running quite slow probably because the texture coordinates
are being computed by the GPU for every frame. This was done because
there are rendering issues with the seam of the texture otherwise.