Corrected Scaling
Added a camera class which can be passed to globe, lines, etc. to manage
the projection and view matrices. The camera now translates to the clip
space correctly so the globe/lines are no longer normalized in their
vertex arrays. This makes it easier to sync things up since everything
is in units of meters. Next step is to plot points.