Compute orbits in a separate thread
This gets rid of the noticable pause when updating. Performance is
acceptable at this time (on a Pi Zero), so I'm not going to worry about
making the calculation any more efficient. The TLE file to read should
probably be an argument to the program. Next up is to look into some
nice graphical effects like a glow for the atmosphere.